Create README.md
Family Game Catalogue

Game
-name:String
-minAge:int
-maxAge:int
-minPlayers:int
-maxPlayers:int
-minPlayTime:int
-maxPlayTime:int
+getName():String
+getMinAge():int
+getMaxAge():int
+getMinPlayers():int
+getMaxPlayers():int
+getMinPlayTime():int
+getMaxPlayTime():int
+setName():void
+setMinAge():void
+setMaxAge():void
+setMinPlayers():void
+setMaxPlayers():void
+setMinPlayTime():void
+setMaxPlayTime():void
+compareName(String):boolean
+compareTo(Game):int
+readGame(Scanner):void
+toString():String

GameCatalogue
-index:SLL<Integer>
-games:SLL<Game>
+GameCatalogue():void
+addGame(Game):void
+getGame(int):Game
+getSize():int
+findGame(String):Game
+readFile(Scanner):void
+removeGame(String):void
+sortCatalogue():void
+toString():String


GameCatalogueDriver
+main(String[]):void


Create a class (Game from above corresponding UML diagram) that handles names of games, the minimum/maximum age of the players, minimum/maximum number of players and the minimum/maximum time to play the game as if you were making a database of your own collection. In addition to the standard accessor and mutator methods from the instance fields, the class will have two different compare methods; a method to read a single game from a file, and a method that will output the data in a user friendly format.
Create a class (GameCatalogue from the above corresponding UML diagram) that maintains the catalogue of games. It should be able to handle an initial population from a file. In addition, the class should be able to add a new game to the catalogue and remove a game identified by name from the catalogue.
Finally, create a class (GameCatalogueDriver from the above corresponding UML diagram) that tests all methods (any method not tested is minus 5 points) from both the Game and GameCatalogue classes.